refactor: streamline compare.py file with changes
There are some major changes to compare.py in this commit.

Firstly, a lot of variables names have been changed in this file and how
this may impact other files has not been tested or remarked in any way.
Secondly, three new functions have been extracted from the previously
named "compare" function and in my humble opinion, the code is a bit
easier to read and understand. Finally, it should be noted that the
jargon associated with this file ("similitude factor") has been removed
for ease of understanding. If the current terms are deemed unjustified
or inaccurate, they should be replaced with appropriate terms. Having
little knowledge of the jargon of this field, I made an attempt at
picking names that were easy to understand for anyone.